+++ packages/orq-rc/src/lib/config.ts
@@ -58,7 +58,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.11.0-rc.3",
-  genVersion: "2.893.0",
-  userAgent: "speakeasy-sdk/typescript 4.11.0-rc.3 2.893.0 2.0 @orq-ai/node",
+  sdkVersion: "4.11.0-rc.4",
+  genVersion: "2.897.1",
+  userAgent: "speakeasy-sdk/typescript 4.11.0-rc.4 2.897.1 2.0 @orq-ai/node",
 } as const;