+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-08T13:07:17.066Z",
+    "2026-06-09T07:16:24.404Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {